try to fix err on ci/cd
