+++ src/main.tsx
@@ -8,6 +8,7 @@ import App from './App';
 import { AudioPlayerProvider } from './contexts/AudioPlayerProvider';
 import './index.css';
 import './material-components.ts';
+import './styles/material-theme/dark.css'; 
 
 const rootElement = document.getElementById('root');
 if (!rootElement) {
